hotfix2 at 06-12

+++ 0x00-python_variable_annotations/6-sum_mixed_list.py
@@ -7,7 +7,7 @@
 from typing import List, Union
 
 
-def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> float:
+def sum_mixed_list(mxd_lst: List[Union[int, float]]) -> float:
     """
     Calculate the sum of a mixed list of floats and integers.
 
@@ -19,4 +19,4 @@ def sum_mixed_list(mxd_lst: List[Union[float, int]]) -> float:
         float: The sum of the elements in the mixed list.
 
     """
-    return sum(mxd_lst)
+    return float(sum(mxd_lst))
